Fall into insanity
Explosion effect added for fun!

Notes to self:
Add "UFO" characters as the mysterious objects to be collect, each grant player 100000000 points and collecting all 3 is the necessary part to win the game
Add "KEY" characters as one of the cutscene, so that the player need to be genius to win the game, after the player collected "UFO" characters they unlock this cutscene. They need to pause the cutscene when it plays and collect the 3 characters

When a player have "UFO" and "KEY", they win the game, and victory cutscene play, awards, achievements, bla bla bla will be granted